+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Other
 
-* Speedup archetype access by 10%, by elimination of bounds checks (!125)
+* Speedup archetype access by 10%, by elimination of bounds checks (!126)
+* Minor optimizations of component storage (!127)
 
 ## [[v0.4.5]](https://github.com/mlange-42/arche/compare/v0.4.4...v0.4.5)